* aprx.h, digipeater.c:     Use 'float' for ratelimit variables.  Permitting as low     rate definitions as 'once per 10 minutes.
* dprsgw.c:
    XOR checksums of "GPS" mode packets.  No Rx-iGate of them yet.

+++ aprx.h
@@ -474,9 +474,9 @@ struct digipeater_source {
 	char		      *via_path; // for APRSIS only
 	uint8_t		       ax25viapath[7]; // APRSIS
 
-	int		       tokenbucket;
-	int		       tbf_increment;
-	int		       tbf_limit;
+	float		       tokenbucket;
+	float		       tbf_increment;
+	float		       tbf_limit;
 
 	// Viscous queue is at <source>, but used dupechecker
 	// is <digipeater> -wide, common to all sources in that
@@ -501,9 +501,9 @@ struct digipeater_source {
 
 struct digipeater {
 	struct aprx_interface *transmitter;
-	int		       tokenbucket;
-	int		       tbf_increment;
-	int		       tbf_limit;
+	float		       tokenbucket;
+	float		       tbf_increment;
+	float		       tbf_limit;
 
 	dupecheck_t           *dupechecker;
 	historydb_t	      *historydb;
